feat: move streak logic to streak page and only check previous session in SessionControl
